Move flashmem from hostlib into spi_flash Component
